[#684] Merge Reader Dashboard into Profile Portfolio tab
- Portfolio tab now includes paginated Trading History with buy/sell badges,
  storyline titles, USD values — visible for all profiles (public on-chain data)
- Donation history (given as reader) gated to own profile only
- /reader now redirects to /profile/[address]?tab=portfolio
